+++ src/Routing/PhrouteRouter.php
@@ -2,6 +2,8 @@
 
 namespace Rougin\Slytherin\Routing;
 
+use Phroute\Phroute\RouteCollector;
+
 /**
  * Phroute Router
  *
@@ -15,4 +17,78 @@ namespace Rougin\Slytherin\Routing;
  */
 class PhrouteRouter extends Router
 {
+    /**
+     * @var \Phroute\Phroute\RouteCollector
+     */
+    protected $collector;
+
+    /**
+     * Initializes the router instance.
+     *
+     * @param array<int, array<int, \Interop\Http\ServerMiddleware\MiddlewareInterface[]|string[]|string>> $routes
+     */
+    public function __construct(array $routes = array())
+    {
+        $this->collector = new RouteCollector;
+
+        parent::__construct($routes);
+    }
+
+    /**
+     * Adds a new raw route.
+     *
+     * @param  string                                                        $method
+     * @param  string                                                        $uri
+     * @param  callable|string[]|string                                      $handler
+     * @param  \Interop\Http\ServerMiddleware\MiddlewareInterface[]|string[] $middlewares
+     * @return self
+     */
+    public function add($method, $uri, $handler, $middlewares = array())
+    {
+        parent::add($method, $uri, $handler, $middlewares);
+
+        // Returns the recently added route from parent ---
+        $route = $this->routes[count($this->routes) - 1];
+        // ------------------------------------------------
+
+        $this->collector->addRoute($method, $route->getUri(), $route->getHandler());
+
+        return $this;
+    }
+
+    /**
+     * @param  \Rougin\Slytherin\Routing\RouteInterface[] $routes
+     * @return \Phroute\Phroute\RouteDataArray
+     */
+    public function asParsed(array $routes)
+    {
+        foreach ($routes as $route)
+        {
+            $this->collector->addRoute($route->getMethod(), $route->getUri(), $route->getHandler());
+        }
+
+        return $this->collector->getData();
+    }
+
+    /**
+     * Returns a listing of parsed routes.
+     *
+     * @param  \Rougin\Slytherin\Routing\RouteInterface[] $routes
+     * @return mixed|null
+     */
+    public function parsed(array $routes = array())
+    {
+        return $this->collector->getData();
+    }
+
+    /**
+     * @param  \Phroute\Phroute\RouteCollector $collector
+     * @return self
+     */
+    public function setCollector(RouteCollector $collector)
+    {
+        $this->collector = $collector;
+
+        return $this;
+    }
 }