+++ tests/Routing/DispatcherTestCases.php
@@ -3,6 +3,7 @@
 namespace Rougin\Slytherin\Routing;
 
 use Rougin\Slytherin\Fixture\Classes\NewClass;
+use Rougin\Slytherin\Testcase;
 
 /**
  * Dispatcher Test Cases
@@ -10,7 +11,7 @@ use Rougin\Slytherin\Fixture\Classes\NewClass;
  * @package Slytherin
  * @author  Rougin Gutib <rougingutib@gmail.com>
  */
-class DispatcherTestCases extends \Rougin\Slytherin\Testcase
+class DispatcherTestCases extends Testcase
 {
     /**
      * @var \Rougin\Slytherin\Routing\DispatcherInterface
@@ -34,11 +35,30 @@ class DispatcherTestCases extends \Rougin\Slytherin\Testcase
 
         $router->post('/', 'NewClass@store');
 
-        $router->get('/hi', function () {
+        $router->get('/hi', function ()
+        {
             return 'Hi and this is a callback';
         });
 
         $this->dispatcher = new Dispatcher($router);
+    }
+
+    /**
+     * Tests DispatcherInterface::dispatch with callback.
+     *
+     * @return void
+     */
+    public function testDispatchMethodWithCallback()
+    {
+        $this->exists(get_class($this->dispatcher));
+
+        $route = $this->dispatcher->dispatch('GET', '/hi');
+
+        $expected = 'Hi and this is a callback';
+
+        $actual = $this->resolve($route);
+
+        $this->assertEquals($expected, $actual);
     }
 
     /**
@@ -52,13 +72,13 @@ class DispatcherTestCases extends \Rougin\Slytherin\Testcase
 
         $controller = new NewClass;
 
-        list($function) = $this->dispatcher->dispatch('GET', '/');
+        $route = $this->dispatcher->dispatch('GET', '/');
 
-        $expected = (string) $controller->index();
+        $expected = $controller->index();
 
-        $result = $this->result($function);
+        $actual = $this->resolve($route);
 
-        $this->assertEquals($expected, $result);
+        $this->assertEquals($expected, $actual);
     }
 
     /**
@@ -72,31 +92,13 @@ class DispatcherTestCases extends \Rougin\Slytherin\Testcase
 
         $controller = new NewClass;
 
-        list($function) = $this->dispatcher->dispatch('POST', '/');
+        $route = $this->dispatcher->dispatch('POST', '/');
 
-        $expected = (string) $controller->store();
+        $expected = $controller->store();
 
-        $result = $this->result($function);
+        $actual = $this->resolve($route);
 
-        $this->assertEquals($expected, $result);
-    }
-
-    /**
-     * Tests DispatcherInterface::dispatch with callback.
-     *
-     * @return void
-     */
-    public function testDispatchMethodWithCallback()
-    {
-        $this->exists(get_class($this->dispatcher));
-
-        list($function) = $this->dispatcher->dispatch('GET', '/hi');
-
-        $expected = 'Hi and this is a callback';
-
-        $result = $this->result($function);
-
-        $this->assertEquals($expected, $result);
+        $this->assertEquals($expected, $actual);
     }
 
     /**
@@ -106,7 +108,7 @@ class DispatcherTestCases extends \Rougin\Slytherin\Testcase
      */
     public function testDispatchMethodWithError()
     {
-        $this->setExpectedException('UnexpectedValueException');
+        $this->setExpectedException('BadMethodCallException');
 
         $this->exists(get_class($this->dispatcher));
 
@@ -120,34 +122,11 @@ class DispatcherTestCases extends \Rougin\Slytherin\Testcase
      */
     public function testDispatchMethodWithInvalidMethod()
     {
-        $this->setExpectedException('UnexpectedValueException');
+        $this->setExpectedException('BadMethodCallException');
 
         $this->exists(get_class($this->dispatcher));
 
         $this->dispatcher->dispatch('TEST', (string) '/');
-    }
-
-    /**
-     * Returns result from the dispatched route.
-     *
-     * @param  array $function
-     * @return mixed
-     */
-    protected function result($function)
-    {
-        if (is_array($function) === true) {
-            list($callback, $parameters) = $function;
-
-            if (is_array($callback) === true) {
-                list($class, $method) = $callback;
-
-                $callback = array(new $class, $method);
-            }
-
-            return call_user_func_array($callback, $parameters);
-        }
-
-        return $function;
     }
 
     /**
@@ -158,7 +137,8 @@ class DispatcherTestCases extends \Rougin\Slytherin\Testcase
      */
     protected function exists($dispatcher)
     {
-        if ($dispatcher === 'Rougin\Slytherin\Routing\FastRouteDispatcher') {
+        if ($dispatcher === 'Rougin\Slytherin\Routing\FastRouteDispatcher')
+        {
             $exists = interface_exists('FastRoute\Dispatcher');
 
             $message = (string) 'FastRoute is not installed.';
@@ -166,7 +146,8 @@ class DispatcherTestCases extends \Rougin\Slytherin\Testcase
             $exists || $this->markTestSkipped((string) $message);
         }
 
-        if ($dispatcher === 'Rougin\Slytherin\Routing\PhrouteDispatcher') {
+        if ($dispatcher === 'Rougin\Slytherin\Routing\PhrouteDispatcher')
+        {
             $exists = class_exists('Phroute\Phroute\Dispatcher');
 
             $message = (string) 'Phroute is not installed.';
@@ -174,4 +155,27 @@ class DispatcherTestCases extends \Rougin\Slytherin\Testcase
             $exists || $this->markTestSkipped((string) $message);
         }
     }
+
+    /**
+     * Returns result from the dispatched route.
+     *
+     * @param  \Rougin\Slytherin\Routing\RouteInterface $route
+     * @return mixed
+     */
+    protected function resolve(RouteInterface $route)
+    {
+        $handler = $route->getHandler();
+
+        if (is_array($handler))
+        {
+            list($class, $method) = $handler;
+
+            $handler = array(new $class, $method);
+        }
+
+        $params = $route->getParams();
+
+        return call_user_func_array($handler, $params);
+    }
+
 }